HORUS AI/pyproject.toml
[project]
name = "horus-ai"
version = "0.1.0"
description = "Antifragile video generation orchestrator with provider chains and fallbacks"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "pydantic>=2.5",
  "httpx>=0.27",
  "typer>=0.12",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-mock>=3.11",
  "ruff>=0.5",
  "black>=24.3",
  "mypy>=1.7"
]

[project.scripts]
vgen = "cli.main:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "cli*"]

[tool.black]
line-length = 100

target-version = ["py310"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
